+++ packages/cx-elements/src/components/Spacing/tokens/cxSpacing.ts
@@ -30,4 +30,6 @@ const meta = {
 export default asTokenGroup(meta)({
   WithSiteMargin: 'mx-site-percent md:mx-md-site-percent xl:px-40',
   WithSiteXLConstraint: 'xl:container xl:mx-auto',
+  Gutter: 'p-1 md:p-2 lg:p-3',
+  GutterOffset: '-mx-1 md:-mx-2 lg:-mx-3',
 });