+++ sites/__cxstarter__/gatsby-config.js
@@ -116,5 +116,6 @@ module.exports = {
   flags: {
     DEV_SSR: false,
   },
+  pathPrefix: process.env.GATSBY_PATH_PREFIX || '',
   plugins,
 };